+++ apisix/admin/routes.lua
@@ -20,7 +20,6 @@ local upstreams = require("apisix.admin.upstreams")
 local tostring = tostring
 local type = type
 local loadstring = loadstring
-local table_util = require("apisix.utils.table-util")
 
 
 local _M = {
@@ -228,7 +227,7 @@ function _M.patch(id, conf, args)
 
     local node_value = res_old.body.node.value
 
-    node_value = table_util.merge(node_value, conf);
+    node_value = core.table.merge(node_value, conf);
 
     core.log.info("new conf: ", core.json.delay_encode(node_value, true))
